rename of Endpoint and import statement change
